move check functions to reports
