+++ specparam/modes/__init__.py
@@ -1,5 +1,8 @@
 """Functionality related to definining fit modes."""
 
 from .mode import Mode
-from .check import check_modes
 from .paramdef import ParamDefinition
+
+# Link in report function to list available modes
+from specparam.reports.options import check_modes
+
