+++ specparam/params/__init__.py
@@ -1,4 +1,7 @@
 """Sub-module for functionality related to parameters and parameter conversions."""
 
 from .converter import AperiodicParamConverter, PeriodicParamConverter
-from .check import check_converters
+
+# Link in report function to list available parameter converters
+from specparam.reports.options import check_converters
+
