--- specparam/tests/algorithms/test_check.py
@@ -1,10 +1,0 @@
-"""Tests for specparam.algorithms.check."""
-
-from specparam.algorithms.check import *
-
-###################################################################################################
-###################################################################################################
-
-def test_check_algorithms():
-
-    check_algorithms()
