--- specparam/tests/metrics/test_check.py
@@ -1,10 +1,0 @@
-"""Test functions for specparam.metrics.check."""
-
-from specparam.metrics.check import *
-
-###################################################################################################
-###################################################################################################
-
-def test_check_metrics():
-
-    check_metrics()
